+++ bin/commands/index.js
@@ -1,0 +1,8 @@
+function command (args, flags, context) {
+  context.help()
+}
+
+const args = []
+const flags = []
+
+module.exports = { command, args, flags }